refactor: extract editor context logic to kilocode/editor-context.ts
Move formatTime and editor context env-line building into a dedicated
kilocode-specific file to minimize changes in the shared system.ts.
The only changes to system.ts are now: import, signature, and one
spread call.
